requirements.txt
PyQt5==5.15.11
google-generativeai==0.8.6
requests==2.33.1
pillow==12.1.1
pywin32==311
winrt-runtime==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1
winrt-Windows.Foundation.Numerics==3.2.1
winrt-Windows.Globalization==3.2.1
winrt-Windows.Graphics.Imaging==3.2.1
winrt-Windows.Media.Ocr==3.2.1
winrt-Windows.Storage.Streams==3.2.1
